Crate trx_rs

trx-rs

Rust library for reading and writing TRX tractography files.

TRX is a file format for storing brain tractography streamlines as memory-mapped binary arrays, either as plain directories or ZIP archives.

Re-exports

pub use any_trx_file::AnyTrxFile;

pub use any_trx_file::PositionsRef;

pub use dtype::DType;

pub use dtype::TrxScalar;

pub use error::Result;

pub use error::TrxError;

pub use formats::convert;

pub use formats::detect_format;

pub use formats::inspect_vtk_declared_space;

pub use formats::read_tractogram;

pub use formats::vtk_import_warnings;

pub use formats::write_tractogram;

pub use formats::ConversionOptions;

pub use formats::Format;

pub use formats::VtkCoordinateMode;

pub use formats::VtkCoordinateSpace;

pub use header::Header;

pub use mmap_backing::MmapBacking;

pub use ops::build_streamline_aabbs;

pub use ops::build_streamline_aabbs_from_slices;

pub use ops::concatenate_any_trx;

pub use ops::copy_metadata;

pub use ops::copy_metadata;

pub use ops::copy_metadata_any_trx;

pub use ops::difference;

pub use ops::difference_indices;

pub use ops::intersection;

pub use ops::intersection_indices;

pub use ops::merge_trx_shards;

pub use ops::query_aabb;

pub use ops::query_aabb_cached;

pub use ops::remove_duplicates;

pub use ops::remove_duplicates_tractogram;

pub use ops::retain_representative_indices;

pub use ops::retain_tractogram_representative_indices;

pub use ops::streamline_union;

pub use ops::subset_streamlines;

pub use ops::ConcatenateOptions;

pub use ops::CopyMetadataOptions;

pub use ops::DuplicateRemovalMode;

pub use ops::DuplicateRemovalParams;

pub use ops::StreamlineAabb;

pub use reference::header_from_reference;

pub use stream::TrxStream;

pub use tractogram::Tractogram;

pub use transform::apply_transform;

pub use transform::apply_transform_in_place;

pub use trx_file::DataArray;

pub use trx_file::DataArrayInfo;

pub use trx_file::TrxFile;

pub use typed_view::TypedView2D;

pub use vertex::Position3;

pub use fit::fit_catmull_rom_indices;

pub use fit::sample_catmull_rom;

pub use fit::sample_catmull_rom_into;

pub use fit::simplify_streamline;

pub use fit::simplify_tractogram;

pub use fit::FittedMarker;

pub use fit::SimplifyOptions;

pub use fit::SimplifyStats;

pub use fit::FITTED_MARKER_KEY;

Modules

any_trx_file

dtype

error

fit

Polyline → Catmull-Rom fitting.

formats

header

io

mmap_backing

ops

Operations on TRX tractogram data: merge, subset, deduplicate, query, and copy metadata.

reference

stream

tractogram

transform

Apply spatial transforms to streamline coordinates.

trx_file

typed_view

vertex