trx_rs

Module formats

Re-exports

pub use vtk::inspect_vtk_declared_space;

pub use vtk::vtk_import_warnings;

pub use vtk::VtkCoordinateMode;

pub use vtk::VtkCoordinateSpace;

Modules

tck

trk

tt

vtk

Structs

ConversionOptions

Options that control tractogram conversion behaviour.

Enums

Format

Supported tractogram file formats.

Functions

convert

Convert between tractogram file formats in one step.

detect_format

Detect the tractogram format from a file path or directory.

read_tractogram

Read a tractogram from any supported format into the neutral in-memory representation.

write_tractogram

Write a tractogram to any supported output format.